+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - `Innmind\OperatingSystem\Sockets` is now a final class
 - `Innmind\OperatingSystem\Remote` is now a final class
 - `Innmind\OperatingSystem\Ports` is now a final class
+- `Innmind\OperatingSystem\Filesystem` is now a final class
 
 ### Fixed